requirements.txt

matplotlib==3.5.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pefile==2023.2.7
pipenv==2022.10.12
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pymongo==4.6.1
pyparsing==3.0.9
python-amazon-sp-api==0.17.5
python-editor==1.0.4
python-dotenv==0.21.0
selenium==4.11.2
geckodriver-autoinstaller==0.1.0
marionette-driver==3.4.0
selenium-firefox==2.0.8
webdriver-manager==4.0.2
beautifulsoup4

